#155 - * chore: split middleware into new file
